Moved qs validation to list action. Don't overwrite item or items in load stage
