+++ src/actions/create.js
@@ -3,16 +3,20 @@ const { generate: uuid } = require('short-uuid');
 async function create(ev, ctx) {
   const { body, session } = ev;
 
-  ev.item = {
-    ...body,
-    id: body.id || uuid(),
-    accountId: session.accountId,
-    version: 1,
-    createdAt: new Date().toISOString(),
-    updatedAt: new Date().toISOString(),
-    createdBy: session.email,
-    updatedBy: session.email
-  };
+  await this.withHooks('load', async() => {
+    if (ev.item) return;
+
+    ev.item = {
+      ...body,
+      id: body.id || uuid(),
+      accountId: session.accountId,
+      version: 1,
+      createdAt: new Date().toISOString(),
+      updatedAt: new Date().toISOString(),
+      createdBy: session.email,
+      updatedBy: session.email
+    };
+  });
 
   await this.withHooks('validate', async() => {
     await this.validate(ev.item);
