+++ src/collection/collection.js
@@ -1,5 +1,4 @@
 const pluralize = require('pluralize');
-const Ajv = require('ajv');
 
 /*
  * Base class for collection providers
@@ -32,14 +31,6 @@ class Collection {
   }
 
   async findByQueryString(query, baseFilter) {
-    const isValid = validateQueryString(query);
-
-    if (!isValid) {
-      const err = new Error('validation failed');
-      err.body = validateQueryString.errors;
-      err.status = 422;
-      throw err;
-    }
   }
 }
 
@@ -48,65 +39,3 @@ function kebabize(str) {
 }
 
 module.exports = Collection;
-
-const ajv = new Ajv({
-  allErrors: true,
-  removeAdditional: 'all',
-  useDefaults: true,
-  coerceTypes: true
-});
-
-const validateQueryString = ajv.compile({
-  type: 'object',
-  properties: {
-    q: {
-      type: 'object',
-      default: {}
-    },
-    qor: {
-      type: 'integer',
-      default: 0,
-      enum: [
-        0,
-        1
-      ]
-    },
-    qre: {
-      type: 'integer',
-      default: 0,
-      enum: [
-        0,
-        1
-      ]
-    },
-    qci: {
-      type: 'integer',
-      default: 0,
-      enum: [
-        0,
-        1
-      ]
-    },
-    sort: {
-      type: 'string',
-      default: 'createdAt desc'
-    },
-    pick: {
-      type: 'array',
-      items: {
-        type: 'string'
-      }
-    },
-    limit: {
-      type: 'integer',
-      default: '10',
-      minimum: 0,
-      maximum: 500
-    },
-    offset: {
-      type: 'integer',
-      default: 0,
-      minimum: 0
-    }
-  }
-});
